feat: show audit finding count
